crawler/shopee/requirements.txt
requests==2.31.0
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.0
python-dotenv==1.0.1
fake-useragent==1.5.1
colorlog==6.8.2
playwright==1.42.0
playwright-stealth==1.0.6
